Fractal algorithm finished and working

+++ Mandelbrot.h
@@ -3,6 +3,6 @@ class Mandelbrot
 {
 public:
 	static const int MAX_ITERATIONS = 1000;
-	static int GetIterations(int x, int y);
+	static int GetIterations(double x, double y);
 };
